create rest of simple messages & refactor tests in to common trait

+++ src/main/scala/wire/message/Choke.scala
@@ -2,14 +2,13 @@ package wire.message
 
 import akka.util.ByteString
 
-//TODO: add tests for Choke, Choke comp. object, and thereby SimpleMessage & SimpleMessageValidator
 case class Choke() extends SimpleMessage {
   val serialize: ByteString = serializeWithMessageType(0)
 }
 
 object Choke extends MessageParser with SimpleMessageValidator {
   def apply(str: ByteString): Choke = {
-    checkValidForMessageType(str, 0.toByte, "Choke")
+    checkValidForMessageType(str, 0, "Choke")
     Choke()
   }
 }
